+++ axum-extra/src/extract/mod.rs
@@ -24,9 +24,9 @@ pub mod multipart;
 #[cfg(feature = "scheme")]
 mod scheme;
 
-pub use self::{
-    cached::Cached, host::Host, optional_path::OptionalPath, with_rejection::WithRejection,
-};
+#[allow(deprecated)]
+pub use self::optional_path::OptionalPath;
+pub use self::{cached::Cached, host::Host, with_rejection::WithRejection};
 
 #[cfg(feature = "cookie")]
 pub use self::cookie::CookieJar;
@@ -41,7 +41,10 @@ pub use self::cookie::SignedCookieJar;
 pub use self::form::{Form, FormRejection};
 
 #[cfg(feature = "query")]
-pub use self::query::{OptionalQuery, OptionalQueryRejection, Query, QueryRejection};
+#[allow(deprecated)]
+pub use self::query::OptionalQuery;
+#[cfg(feature = "query")]
+pub use self::query::{OptionalQueryRejection, Query, QueryRejection};
 
 #[cfg(feature = "multipart")]
 pub use self::multipart::Multipart;