+++ axum-macros/tests/typed_path/pass/result_handler.rs
@@ -8,8 +8,6 @@ struct UsersShow {
     id: String,
 }
 
-async fn option_handler(_: Option<UsersShow>) {}
-
 async fn result_handler(_: Result<UsersShow, PathRejection>) {}
 
 #[derive(TypedPath, Deserialize)]
@@ -20,7 +18,6 @@ async fn result_handler_unit_struct(_: Result<UsersIndex, StatusCode>) {}
 
 fn main() {
     _ = axum::Router::<()>::new()
-        .typed_get(option_handler)
         .typed_post(result_handler)
         .typed_post(result_handler_unit_struct);
 }