+++ axum-core/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# Unreleased
+
+- **breaking:**: `Option<T>` as an extractor now requires `T` to implement the
+  new trait `OptionalFromRequest` (if used as the last extractor) or
+  `OptionalFromRequestParts` (other extractors) ([#2475])
+
+[#2475]: https://github.com/tokio-rs/axum/pull/2475
+
 # 0.5.0
 
 ## alpha.1