+++ axum-extra/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning].
 
 # Unreleased
 
+- **breaking:** `Option<Query<T>>` no longer swallows all error conditions, instead rejecting the
+  request in many cases; see its documentation for details ([#2475])
+- **changed:** Deprecated `OptionalPath<T>` and `OptionalQuery<T>` ([#2475])
 - **fixed:** `Host` extractor includes port number when parsing authority ([#2242])
 - **changed:** The `multipart` feature is no longer on by default ([#3058])
 - **added:** Add `RouterExt::typed_connect` ([#2961])
@@ -16,6 +19,7 @@ and this project adheres to [Semantic Versioning].
 - **added:** Add `FileStream` for easy construction of file stream responses ([#3047])
 
 [#2242]: https://github.com/tokio-rs/axum/pull/2242
+[#2475]: https://github.com/tokio-rs/axum/pull/2475
 [#3058]: https://github.com/tokio-rs/axum/pull/3058
 [#2961]: https://github.com/tokio-rs/axum/pull/2961
 [#2962]: https://github.com/tokio-rs/axum/pull/2962